Fix loading wavs for Android.

+++ androidaudio/androidsoundeffect.h
@@ -39,7 +39,7 @@ public slots:
 
 private:
     char* mBuffer;
-    off_t mLength;
+    int mLength;
 
     QString mPath;
